refactor: move appending synced status to claims facade
